Moved FormatDate class
